requirements.txt
fastapi~=0.115.12
uvicorn[standard]
pydantic[email]~=2.11.4
motor~=3.7.0
python-multipart
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.4.0
pymongo~=4.12.1
python-dotenv~=1.1.0
pydantic-settings~=2.9.1
starlette~=0.46.2
fitz~=0.0.1.dev2
python-docx~=1.1.2
XlsxWriter
pymupdf
python-pptx~=1.0.2
groq
docx~=0.2.4
pathlib~=1.0.1
requests~=2.32.3